This is a further worked on notebook Synthetic mock community analysis.ipynb being renomed without spaces.
I fixed some of the scoring algorithms in the following functions:

* mapping_results
* pull_mapping_results

These scoring functions seem to work well with the long ITS sequences. I also developed some more functions that summarize the results at all taxonomics
ranks. This works for all. At times Tavish will have to check the taxonomics assignments of at the family level for the queries. There seems to be a bit
of inconsistency in Qiime and the homemade database. I think Qiime isn't great here.

One idea would be to pull in the Qiime assignemnt when possible and compare this as well.

Overall assiging against Qiime doesn't work great. Will have to explore maybe a bit more the filtering of results including the nmatch of minimap2.
Currently a bit unsure if this is fully correct. Maybe another algorithm would work better here. Good to mention in the discussion.